omega/analyzer/worker/tools/requirements.txt
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==38.0.1
fetchcode==0.1.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
packageurl-python==0.9.9
pycparser==2.21
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
urllib3==1.26.9
